mysite/requirements.txt
# Dev environment
pip
autopep8

# App
asgiref==3.7.2
Django==5.0.13
sqlparse==0.4.4
typing_extensions==4.9.0
Pillow==11.3.0
django-impersonate==1.9.5
